+++ CHANGELOG.md
@@ -8,8 +8,9 @@ Log of unreleased changes.
 
 ### Changed
 
-- `*_liter` and `*_unit` prices are now calculated.
+- `*_liter` and `*_unit` prices are now calculated fields.
 - Deploying to `stage` does not require a tag.
+- Prices are stored as `decimal` instead of `integer`.
 
 ### Fixes